% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchorpartner15.R
\name{anchorpartner15}
\alias{anchorpartner15}
\title{Plausible values imputation within couples using the PIAAC-L 2015 assessment data in the domains of literacy, numeracy, reading and mathematics}
\usage{
anchorpartner15(path, Xanchor = NULL, Xpartner = NULL, nopvs = 10,
  itermcmc = 22000, burnin = 2000)
}
\arguments{
\item{path}{full path of the folder containing the data files ZA5845 and ZA5989_Persons_15 in Stata format.}

\item{Xanchor}{data frame containing the permanent person ID (named \code{pnrfestid}) and background variable from the PIAAC and PIAAC-L
Scientific Use Files. They can be numeric or factor variables and contain missing values coded as \code{NA}. With \code{Xanchor} set to \code{NULL} (default) no background variables of the anchor person are considered. With both \code{Xanchor} and \code{Xpartner} set to \code{NULL} an empty population model will be estimated.}

\item{Xpartner}{data frame containing the permanent person ID (named \code{pnrfestid}) and background variable from the PIAAC and PIAAC-L
Scientific Use Files. They can be numeric or factor variables and contain missing values coded as \code{NA}. With \code{Xpartner} set to \code{NULL} (default) no background variables of the partner are considered. With both \code{Xanchor} and \code{Xpartner} set to \code{NULL} an empty population model will be estimated.}

\item{nopvs}{number of plausible values to draw for each respondent.}

\item{itermcmc}{number of MCMC iterations.}

\item{burnin}{number of burnin iterations.}
}
\value{
list with \code{nopvs} elements, each containing a data frame of the sequential ID, plausible values for
each dimension and, if specified, imputed versions of \code{X}. Additionally each list element is saved as a 
Stata file in the folder specified by \code{path}.
}
\description{
This function estimates a four-dimensional latent regression model (first dimension: PIAAC-L 2015 literacy, second dimension:
PIAAC-L 2015 numeracy, third dimension: PIAAC-L 2015 reading, fourth dimension: PIAAC-L 2015 mathematics) for binary item response data considering
partially missing covariate data. Anchor persons and their partners are treated as one observation. For more detailed information on the statistical model and the estimation algorithm, see the
PIAAC-L technical report on scaling (Carstensen, Gaasch & Rothaug, 2017).
}
\references{
Carstensen, C. H., Gaasch, J.-C., & Rothaug, E. (2017). Scaling PIAAC-L cognitive data: technical report.
Manuscript in preparation.
}
