% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litnum1215.R
\name{litnum1215}
\alias{litnum1215}
\title{Plausible values imputation using the PIAAC 2012 and PIAAC-L 2015 assessment data in the domains of literacy and numeracy}
\usage{
litnum1215(path, X = NULL, nopvs = 10, itermcmc = 22000, burnin = 2000)
}
\arguments{
\item{path}{full path of the folder containing the data files ZA5845 and ZA5989_Persons_15 in Stata format.}

\item{X}{data frame containing the sequential ID (named \code{seqid}) and background variable from the PIAAC and PIAAC-L
Scientific Use Files. They can be numeric or factor variables and contain missing values coded as \code{NA}. With \code{X} set to \code{NULL} (default) an empty population model is estimated.}

\item{nopvs}{number of plausible values to draw for each respondent.}

\item{itermcmc}{number of MCMC iterations.}

\item{burnin}{number of burnin iterations.}
}
\value{
list with \code{nopvs} elements, each containing a data frame of the sequential ID, plausible values for
each dimension and, if specified, imputed versions of \code{X}. Resulting plausible values are transformed onto the PIAAC 2012 scale (weighted means and standard deviations based on the SUF). Additionally each list element is saved as a 
Stata file in the folder specified by \code{path}.
}
\description{
This function estimates a four-dimensional latent regression item response model (first dimension: PIAAC 2012 literacy,
second dimension: PIAAC 2012 numeracy, third dimension: PIAAC-L 2015 literacy, fourth dimension: PIAAC-L 2015 numeracy)
for binary item response data considering partially missing covariate data. For more detailed information on the statistical
model and the estimation algorithm, see the PIAAC-L technical report on scaling (Carstensen, Gaasch & Rothaug, 2017).
}
\references{
Carstensen, C. H., Gaasch, J.-C., & Rothaug, E. (2017). Scaling PIAAC-L cognitive data: technical report.
Manuscript in preparation.
}
