% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litnumps12.R
\name{litnumps12}
\alias{litnumps12}
\title{Plausible values imputation using the PIAAC 2012 assessment data in the domains of literacy, numeracy and problem solving}
\usage{
litnumps12(path, XplusVars = NULL, XplusVarsfactor = NULL, nopvs = 10,
  itermcmc = 22000, burnin = 2000)
}
\arguments{
\item{path}{full path of the folder containing the following PIAAC and PIAAC-L Scientific Use Files in Stata format: ZA5845,
ZA5989_Persons_14, ZA5989_Household_14, ZA5989_Persons_15, ZA5989_Persons_16 and ZA5989_Household_16.}

\item{XplusVars}{character vector containing additional background variable names from the PIAAC and PIAAC-L
Scientific Use Files ZA5845, ZA5989_Persons_14, ZA5989_Household_14, ZA5989_Persons_15, ZA5989_Persons_16 and
ZA5989_Household_16. With \code{XplusVars} set to \code{NULL} (default) the basic specification of the population model is
estimated.}

\item{XplusVarsfactor}{logical vector of the same length as \code{XplusVars} indicating which elements of \code{XplusVars}
are factor variables.}

\item{nopvs}{number of plausible values to draw for each respondent.}

\item{itermcmc}{number of MCMC iterations.}

\item{burnin}{number of burnin iterations.}
}
\value{
list with \code{nopvs} elements, each containing a data frame of the sequential ID, plausible values for
each dimension and imputed versions of the partially missing covariate data. Additionally each list element is saved as a 
Stata file in the folder specified by \code{path}.
}
\description{
This function estimates a three-dimensional latent regression model (first dimension: PIAAC 2012 literacy, second dimension:
PIAAC 2012 numeracy, third dimension: PIAAC 2012 problem solving) for binary and ordinal item response data considering
partially missing covariate data. For more detailed information on the statistical model and the estimation algorithm, see the
PIAAC-L technical report on scaling (Carstensen, Gaasch & Rothaug, 2017).
}
\references{
Carstensen, C. H., Gaasch, J.-C., & Rothaug, E. (2017). Scaling PIAAC-L cognitive data: technical report.
Manuscript in preparation.
}
